+++ application/views/articles/index.php
@@ -35,8 +35,8 @@
         <div class="article-box">
           <div class="img-container">
             <a href="<?php echo $ac['link']; ?>">
-                <img src="<?php echo base_url(array('uploads', $ac['image_path'])); ?>" onerror="this.src = '<?php echo base_url('assets/icons/default.jpg'); ?>';" class="article-img" alt="<?php echo $ac['title']; ?>">
-              </a>
+							<img src="<?php echo base_url(array('uploads', $ac['image_path'])); ?>" onerror="this.src = '<?php echo base_url('assets/icons/default.jpg'); ?>';" class="article-img" alt="<?php echo $ac['title']; ?>">
+						</a>
             <a class="category-label <?php echo $ac['subcat_slug'];?>" href="<?php echo site_url($ac['subcat_slug']);?>"><?php echo $ac['subcat_name'];?></a>
           </div>
           <div class="article-text-section">
@@ -54,7 +54,6 @@
       <?php endif;
     endforeach; ?>
 </div>
-
 
 <nav>
 	<div class="article-list-pager">