+++ application/views/pages/home.php
@@ -4,8 +4,8 @@
         <div class="article-box">
           <div class="img-container">
             <a href="<?php echo $ac['link']; ?>">
-                <img src="<?php echo base_url(array('uploads', $ac['image_path'])); ?>" class="article-img" alt="<?php echo $ac['title']; ?>">
-              </a>
+              <img src="<?php echo base_url(array('uploads', $ac['image_path'])); ?>" onerror="this.src = '<?php echo base_url('assets/icons/default.jpg'); ?>';" class="article-img" alt="<?php echo $ac['title']; ?>">
+            </a>
             <a class="category-label <?php echo $ac['subcat_slug'];?>" href="<?php echo site_url($ac['subcat_slug']);?>"><?php echo $ac['subcat_name'];?></a>
           </div>
           <div class="article-text-section">
@@ -23,3 +23,51 @@
       <?php endif;
     endforeach; ?>
 </div>
+
+<nav>
+	<div class="article-list-pager">
+    <?php
+      $link = site_url();
+			$db = ceil($cnt / $limit);
+			$now = ceil($from / $limit) + 1;
+			$start = $now - 3;
+			$end = $now + 8;
+			$prev = $from - $limit;
+      
+      $elem = '';
+      if($from == 0) {
+        $elem = '<a class="btn-pager btn-disabled';
+      } else {
+        if ($prev < 0) // 0 alá ne menjünk a lapozással
+				  $elem = '<a href="' . site_url();
+        else
+          $elem = '<a href="' . site_url($prev);
+        $elem .=  '" class="btn-pager';
+      }
+			echo $elem . '">Előző</a></button>';
+			
+			echo '<div>';
+      if($start > 1) {
+        echo '<span class="pager-link">...</span>';
+      }
+			for($i = $start; $i <= $end && $i <= $db; ++$i) {
+				if($i > 0) {
+					if($i === $now)
+						echo '<span class="pager-link">' . $i . '</span>';
+					else
+					  echo '<a class="pager-link" href="' . site_url(($i - 1) * $limit) . '">' . $i . '</a>';
+				}
+      }
+      if($db > $end) {
+        echo '<span class="pager-link">...</span>';
+      }
+			echo '</div>';
+
+			if($from+$limit < $cnt) //az összes elemszámot még nem érjük el
+				echo '<a class="btn-pager" href="' . site_url($from + $limit) . '">';
+			else // az összes elemszám fölé ne menjünk
+				echo '<a class="btn-pager btn-disabled href="' . site_url($from) . '">';
+			echo 'Következő</a>';
+		?>
+	</div>
+</nav>