+++ application/logs/CHANGELOG.md
@@ -5,6 +5,7 @@
 - logó csere .svg-re
 - a cikklistákban a címek is linkek
 - cikk törlése gomb eltüntetése 4-es hozzáférési szint alatt
+- főoldali lapozás elkészítése
 
 ## Változások 2017.07.29.